backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0    # PostgreSQL driver
pillow>=9.5.0             # For image & file uploads
python-decouple>=3.8       # For .env file config
drf-yasg>=1.21.0
pytest>=7.3.0            # Testing
pytest-django>=4.5.0